enter postgame when less than 2 snakes alive
