+++ client/src/redux/display/displayHelpers.js
@@ -40,49 +40,3 @@ export const aggregateBoards = (lastTu) => {
 export const getInitialDisplayBoard = () => (
   aggregateBoards(constants.INITIAL_TU)
 );
-
-export const getNextDisplayBoard = () => {
-  const state = store.getState();
-  const tu = state.info.tu;
-  const newBoard = aggregateBoards(tu);
-
-  // combine boards needed for current snake length - 1
-  // (predicted move will make full snake length)
-
-  const snakesObj = helpers.deepClone(state.snakes);
-  const snakeIds = Object.keys(snakesObj);
-
-  const length = snakeHelpers.getSnakeLength(tu);
-
-  let snake;
-  let mostRecentTu;
-  let next;
-
-  // predict next moves, fill in missing data with predictions
-  snakeIds.forEach((id) => {
-    if (snakeHelpers.snakeIsAlive(id)) {
-      snake = snakesObj[id];
-      mostRecentTu = Number(snake.positions.byIndex[0]);
-
-      // run once for all snakes, more for snakes with missing TUs
-      while (mostRecentTu < tu) {
-        // calculate next coordinates (predicted)
-        next = snakeHelpers.calculateNextCoords(snake.direction, snake.positions.byKey[`${mostRecentTu}`]);
-
-        mostRecentTu += 1;
-
-        // add predicted coordinates to cloned snake object
-        snake.positions.byKey[`${mostRecentTu}`] = next;
-        snake.positions.byIndex.unshift(`${mostRecentTu}`);
-
-        // add next position to newBoard if it is within range
-        // (based on target TU and snake length)
-        if (mostRecentTu <= tu && mostRecentTu > tu - length) {
-          addCoordinatesMutate(newBoard, next, snake, id);
-        }
-      }
-    }
-  });
-
-  return newBoard;
-};
